requirements.txt
# MCP + HTTP
mcp>=0.6.0
httpx>=0.27.0
pydantic>=2.7.0
python-dateutil>=2.9.0
dateparser>=1.2.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# CEAT report (ceat-nagpur_test.py)
psycopg2-binary>=2.9.0
reportlab>=4.0.0
pandas>=2.0.0
pytz>=2023.3
